# Refactor - Adjust permissions query for performance