--- a/src/Database/Adapter/MariaDB.php
+++ b/src/Database/Adapter/MariaDB.php
@@ -530,7 +530,8 @@ class MariaDB extends Adapter
 
         $permissions = (Authorization::$status) ? "INNER JOIN {$this->getNamespace()}.{$name}_permissions as table_permissions
             ON table_main._uid = table_permissions._uid
-            AND table_permissions._action = 'read' AND table_permissions._role IN (".implode(',', $roles).")" : ''; // Disable join when no authorization required
+            AND table_permissions._action = 'read'" : ''; // Disable join when no authorization required
+        $permissions2 = (Authorization::$status) ? " AND table_permissions._role IN (".implode(',', $roles).")" : ''; // Disable join when no authorization required
 
         foreach($queries as $i => $query) {
             $conditions = [];
@@ -546,6 +547,7 @@ class MariaDB extends Adapter
         $stmt = $this->getPDO()->prepare("SELECT table_main.* FROM {$this->getNamespace()}.{$name} table_main
             {$permissions}
             WHERE ".implode(' AND ', $where)."
+            {$permissions2}
             GROUP BY table_main._uid 
             {$order}
             LIMIT :offset, :limit;
@@ -603,7 +605,8 @@ class MariaDB extends Adapter
 
         $permissions = (Authorization::$status) ? "INNER JOIN {$this->getNamespace()}.{$name}_permissions as table_permissions
             ON table_main._uid = table_permissions._uid
-            AND table_permissions._action = 'read' AND table_permissions._role IN (".implode(',', $roles).")" : ''; // Disable join when no authorization required
+            AND table_permissions._action = 'read'" : ''; // Disable join when no authorization required
+        $permissions2 = (Authorization::$status) ? " AND table_permissions._role IN (".implode(',', $roles).")" : ''; // Disable join when no authorization required
 
         foreach($queries as $i => $query) {
             $conditions = [];
@@ -617,6 +620,7 @@ class MariaDB extends Adapter
         $stmt = $this->getPDO()->prepare("SELECT COUNT(1) as sum FROM (SELECT 1 FROM {$this->getNamespace()}.{$name} table_main
             {$permissions}
             WHERE ".implode(' AND ', $where)."
+            {$permissions2}
             GROUP BY table_main._uid 
             {$limit}) table_count
         ");